requirements.txt
fastapi==0.115.5
pydantic==2.9.2
pyyaml==6.0
langchain_community==0.3.7
langchain_openai==0.2.8
langchain_core==0.3.18
faiss-cpu==1.9.0
numpy==1.26.4
uvicorn==0.32.0